refactor: Move useTheme into its own file
